+++ src/list-data/index.ts
@@ -6,3 +6,4 @@ export * from './intl-jurisdictions'
 export * from './location-data'
 export * from './request-action-mapping'
 export * from './request-actions-data'
+export * from './us-states'
