+++ tests/unit/applicant-info-1.spec.ts
@@ -1,8 +1,8 @@
 import ApplicantInfo1 from '@/components/common/applicant-info-1.vue'
 import { createLocalVue, shallowMount, mount } from '@vue/test-utils'
 import Vuetify from 'vuetify'
-import $intJurisdictions from '@/store/list-data/intl-jurisdictions'
-import $canJurisdictions from '@/store/list-data/canada-jurisdictions'
+import $intJurisdictions from '@/list-data/intl-jurisdictions'
+import $canJurisdictions from '@/list-data/canada-jurisdictions'
 
 // Prevent the warning "[Vuetify] Unable to locate target [data-app]"
 document.body.setAttribute('data-app', 'true')
