+++ tests/unit/names-capture.spec.ts
@@ -1,6 +1,6 @@
 import NamesCapture from '@/components/common/names-capture.vue'
 import { createLocalVue, shallowMount, mount } from '@vue/test-utils'
-import designations from '@/store/list-data/designations'
+import designations from '@/list-data/designations'
 // import store from '@/store/new-request-module'
 import Vuetify from 'vuetify'
